move total_model_params to utils
